mpegts: improvements to DVB SI processing
I've made the table extraid 64 bits, as some entries require 32+X bits
so a 32-bit field wasn't enough.

Added callback for registering destruction (for dealloc'ing memory
alloc'd for opaque).

Added method of registering desciptor tag handlers, saves duplicating
BAT processing used for OpenTV, which uses 1 custom descriptor.

Improved the table completion detection and reduced level of debug
that could swamp tvhlog queue.

+++ src/input/mpegts/dvb.h
@@ -160,7 +160,7 @@ int dvb_table_end
   (struct mpegts_table *mt, struct mpegts_table_state *st, int sect );
 int dvb_table_begin
   (struct mpegts_table *mt, const uint8_t *ptr, int len,
-   int tableid, int extraid, int minlen,
+   int tableid, uint64_t extraid, int minlen,
    struct mpegts_table_state **st, int *sect, int *last, int *ver);
 
 int dvb_pat_callback
